Initial BLoC based design (#30)
* Shuffle files into different packages for bloc architecture

* Move out state for journeys to BLoC architecture. TODO: troubleshoot why new journeys aren't added on rebuild

* Add folders for pages vs components

* Fix style issues
