--- lib/model/inkstep_model.dart
@@ -1,1 +1,0 @@
-class InkstepModel {}
