+++ lib/models/journey_model.dart
@@ -1,0 +1,6 @@
+class JourneyModel {
+  JourneyModel(this.artistName, this.studioName);
+
+  final String artistName;
+  final String studioName;
+}
