+++ lib/ui/components/artist_selection.dart
