+++ lib/ui/components/logo.dart
