+++ lib/ui/pages/info_gathering.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:inkstep/component/grid_image_picker.dart';
-import 'package:inkstep/component/short_text_input.dart';
-import 'package:inkstep/logo.dart';
+import 'package:inkstep/ui/components/grid_image_picker.dart';
+import 'package:inkstep/ui/components/logo.dart';
+import 'package:inkstep/ui/components/short_text_input.dart';
 
 class InfoScreen extends StatefulWidget {
   @override
