+++ pubspec.yaml
@@ -24,6 +24,9 @@ dependencies:
 
   shared_preferences: ^0.5.0
 
+  bloc: ^0.14.0
+  flutter_bloc: ^0.14.0
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
